Roll all the internal APIs that are used to perform self registration into "PluginSelfRegistrationHost"
